Move sidenav stuff into it's own component.
