+++ containers/ecr-viewer/src/app/view-data/components/Demographics.tsx
@@ -4,7 +4,7 @@ import {
   AccordianH3,
   AccordianDiv,
 } from "../component-utils";
-import { SectionConfig } from "./SectionConfig";
+import { SectionConfig } from "./SideNav";
 
 interface DemographicsProps {
   demographicsData: DisplayData[];
