--- containers/ecr-viewer/src/app/view-data/components/SectionConfig.ts
@@ -1,20 +1,0 @@
-export class SectionConfig {
-  title: string;
-  id: string;
-  subNavItems?: SectionConfig[];
-
-  constructor(title: string, subNavItems?: string[] | SectionConfig[]) {
-    this.title = title;
-    this.id = title.toLowerCase().replace(/\s+/g, "-");
-
-    if (subNavItems) {
-      this.subNavItems = subNavItems.map((item) => {
-        if (typeof item === "string") {
-          return new SectionConfig(item);
-        } else {
-          return item;
-        }
-      });
-    }
-  }
-}
